Update chapter4.md
Update to let users know that we need to call the render method inside the IGameLogic instance or else you would not see the triangle being rendered. Anybody could look at the source code to figure this out but since it is a tutorial a little hand holding does not hurt.